tools/bsl/BSL_GUI_source_code/requirements.txt
altgraph==0.17.4
future==1.0.0
iso8601==2.1.0
packaging==24.0
pefile==2023.2.7
pyinstaller==6.5.0
pyinstaller-hooks-contrib==2024.3
pyserial==3.5
pywin32-ctypes==0.2.2
PyYAML==6.0.1
